+++ CHANGELOG.md
@@ -6,3 +6,4 @@ All notable changes to the Zowe Client Python SDK will be documented in this fil
 
 - BugFix: Validation of zowe.config.json file matching the schema [#192](https://github.com/zowe/zowe-client-python-sdk/issues/192)
 - Feature: Added method to load profile properties from environment variables
+- Feature: Added a CredentialManager class to securely retrieve values from credentials and manage multiple credential entries on Windows [#134](https://github.com/zowe/zowe-client-python-sdk/issues/134)